+++ src/solana/SolanaEngine.ts
@@ -74,6 +74,7 @@ export class SolanaEngine extends CurrencyEngine<
   SolanaTools,
   SafeSolanaWalletInfo
 > {
+  lightMode: boolean
   networkInfo: SolanaNetworkInfo
   base58PublicKey: string
   feePerSignature: string
@@ -90,9 +91,10 @@ export class SolanaEngine extends CurrencyEngine<
     env: PluginEnvironment<SolanaNetworkInfo>,
     tools: SolanaTools,
     walletInfo: SafeSolanaWalletInfo,
-    opts: any // EdgeCurrencyEngineOptions
+    opts: EdgeCurrencyEngineOptions
   ) {
     super(env, tools, walletInfo, opts)
+    this.lightMode = opts.lightMode ?? false
     this.networkInfo = env.networkInfo
     this.chainCode = tools.currencyInfo.currencyCode
     this.fetchCors = getFetchCors(env.io)
@@ -547,9 +549,16 @@ export class SolanaEngine extends CurrencyEngine<
     this.queryBlockhash().catch(() => {})
 
     this.addToLoop('queryBalance', ACCOUNT_POLL_MILLISECONDS).catch(() => {})
-    this.addToLoop('queryTransactions', TRANSACTION_POLL_MILLISECONDS).catch(
-      () => {}
-    )
+    if (this.lightMode) {
+      this.tokenCheckTransactionsStatus[this.currencyInfo.currencyCode] = 1
+      for (const edgeToken of Object.values(this.allTokensMap)) {
+        this.tokenCheckTransactionsStatus[edgeToken.currencyCode] = 1
+      }
+    } else {
+      this.addToLoop('queryTransactions', TRANSACTION_POLL_MILLISECONDS).catch(
+        () => {}
+      )
+    }
     await super.startEngine()
   }